ref: create proper pakcage, use ArgumentsParser
